+++ Backend/FinalDemo/APIService/Controllers/OrderController.cs
@@ -98,7 +98,7 @@ namespace KCSAH.APIServer.Controllers
             {
                 return BadRequest("Mapping to order entity failed.");
             }
-            decimal total = 0; // Khởi tạo giá trị cho total
+            int total = 0; // Khởi tạo giá trị cho total
             foreach (var detail in orderMap.OrderDetails)
             {
                 var product = await GetProductAsync(detail.ProductId);
